Consolidate source tree
Reorganize the source tree to make bpfilter more modular and avoid
building the same source file multiple times.

With bfcli being introduced in 90a6136, src/core sources were used
outside of the bpfilter daemon, some other features of the build process
have also been disabled (or broken) over time.

This change reorganize the sources into:
- src: sources for all the project's byproducts: daemon, library, cli.
  This directory is modular, it contains core, shared, lib, ... Which
  can then be used to build other binaries. For example, core and xlate
  are used to build bfcli.
- tests: mostly unit tests for now. End-to-end tests have been removed
  in a previous commit, as they were unused.
- doc: documentation, unmodified except for the coverage target.

lib and cli have been removed, as they also depend on src/core, so they
are now modules in src.
