+++ src/core/bpf.c
@@ -17,8 +17,8 @@
 #include <unistd.h>
 
 #include "core/logger.h"
+#include "core/opts.h"
 #include "generator/nf.h"
-#include "opts.h"
 #include "shared/helper.h"
 
 #define _bf_ptr_to_u64(ptr) ((unsigned long long)(ptr))
