+++ src/lib/bpfilter.h
@@ -5,8 +5,6 @@
 
 #pragma once
 
-#include <bpfilter/shared/request.h>
-#include <bpfilter/shared/response.h>
 #include <stddef.h>
 
 struct ipt_getinfo;
@@ -14,18 +12,6 @@ struct ipt_get_entries;
 struct ipt_replace;
 struct xt_counters_info;
 struct nlmsghdr;
-
-/**
- * @brief Send a request to the daemon and receive the response.
- *
- * @param request Request to send to the daemon. Caller keep ownership of
- * the request. Can't be NULL.
- * @param response Response received from the daemon. It will be allocated
- * by the function and the caller will be responsible for freeing it. Can't
- * be NULL.
- * @return 0 on success, negative errno value on error.
- */
-int bf_send(const struct bf_request *request, struct bf_response **response);
 
 /**
  * @brief Send iptable's ipt_replace data to bpfilter daemon.
