+++ src/lib/generic.c
