+++ src/shared/front.c
