+++ src/shared/generic.c
