+++ src/shared/generic.h
