+++ src/shared/helper.h
