+++ tests/CMakeLists.txt
@@ -5,12 +5,45 @@ add_subdirectory(unit)
 add_subdirectory(integration)
 add_subdirectory(bpf)
 
+get_target_property(shared_srcs shared SOURCES)
+get_target_property(core_srcs core SOURCES)
+get_target_property(generator_srcs generator SOURCES)
+get_target_property(xlate_srcs xlate SOURCES)
+get_target_property(lib_srcs lib SOURCES)
+get_target_property(daemon_srcs daemon SOURCES)
+
 add_custom_target(lint
-    COMMAND ${CLANG_TIDY_BIN} --config-file=${CMAKE_SOURCE_DIR}/.clang-tidy -p ${CMAKE_BINARY_DIR} ${bpfilter_shared_srcs} ${bpfilter_daemon_srcs} ${bpfilter_library_srcs}
+    COMMAND
+        ${CLANG_TIDY_BIN}
+            --config-file=${CMAKE_SOURCE_DIR}/.clang-tidy
+            -p ${CMAKE_BINARY_DIR}
+            ${shared_srcs}
+            ${core_srcs}
+            ${generator_srcs}
+            ${xlate_srcs}
+            ${lib_srcs}
+            ${daemon_srcs}
+            ${CMAKE_SOURCE_DIR}/src/cli/cli.c
     COMMENT "Run linter on source files"
 )
 
 add_custom_target(checkstyle
-    COMMAND ${CLANG_FORMAT_BIN} --style=file:${CMAKE_SOURCE_DIR}/.clang-format --dry-run ${bpfilter_shared_srcs} ${bpfilter_daemon_srcs} ${bpfilter_library_srcs}
+    COMMAND
+        ${CLANG_FORMAT_BIN}
+            --style=file:${CMAKE_SOURCE_DIR}/.clang-format
+            --dry-run
+            ${shared_srcs}
+            ${core_srcs}
+            ${generator_srcs}
+            ${xlate_srcs}
+            ${lib_srcs}
+            ${daemon_srcs}
+            ${CMAKE_SOURCE_DIR}/src/cli/cli.c
     COMMENT "Check source code formatting"
+)
+
+add_custom_target(test
+    COMMAND $<TARGET_FILE:unit_tests>
+    DEPENDS unit_tests
+    COMMENT "Running tests"
 )
